+++ app/src/main/java/com/example/coronastatistics/CoronaActivity.java
@@ -3,31 +3,33 @@ package com.example.coronastatistics;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
+import com.squareup.okhttp.OkHttpClient;
+import com.squareup.okhttp.Request;
+import com.squareup.okhttp.Response;
+
 import java.io.IOException;
 
 public class CoronaActivity extends AppCompatActivity implements View.OnClickListener {
-    Button btnMap;
+    private Button btnMap;
+    private TextView txtStats;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_corona);
 
-        // Get the Intent that started this activity and extract the string
-        Intent intent = getIntent();
-        String message = intent.getStringExtra(MainActivity.EXTRA_MESSAGE);
-
-        // Capture the layout's TextView and set the string as its text
-        TextView textView = findViewById(R.id.editText);
-        textView.setText(message);
-
         btnMap = (Button) findViewById(R.id.btnMap);
+        btnMap.setText("Google Maps");
         btnMap.setOnClickListener((View.OnClickListener) this);
+
+        new AsyncCoronaTask().execute();
     }
 
 
@@ -37,4 +39,55 @@ public class CoronaActivity extends AppCompatActivity implements View.OnClickLis
         Intent intent = new Intent(this, MapsActivity.class);
         startActivity(intent);
     }
+
+
+    /** Uses AsyncTask to create a task away from the main UI thread (to avoid
+     * a NetworkOnMainThreadException). This task takes a
+     * URL string and uses it to create an HttpUrlConnection.
+     * */
+    private class AsyncCoronaTask extends AsyncTask<String, Void, String> {
+        private String url;
+        private String responseBody;
+        private OkHttpClient client = new OkHttpClient();
+        private Response response = null;
+        private TextView txtStats = (TextView) findViewById(R.id.txtStats);
+
+        @Override
+        protected void onPreExecute() {
+            txtStats.setText("Nichts");
+        }
+
+        @Override
+        protected String doInBackground(String... urls) {
+            Request request = new Request.Builder()
+                    .url("https://covid-19-coronavirus-statistics.p.rapidapi.com/v1/stats?country=Canada")
+                    .get()
+                    .addHeader("x-rapidapi-host", "covid-19-coronavirus-statistics.p.rapidapi.com")
+                    .addHeader("x-rapidapi-key", "7e6ace8a6dmsh86c04f7b84eada4p15df6fjsn9c141732c6b9")
+                    .build();
+
+            try {
+                response = client.newCall(request).execute();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            Log.e("Corona", String.valueOf(response.isSuccessful()));
+            try {
+                responseBody = response.body().string();
+                Log.e("Corona", String.valueOf(responseBody));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            return response.toString();
+        }
+        ;
+        @Override
+        protected void onPostExecute(String result) {
+            txtStats.setText(responseBody);
+            Log.e("Henlo", "Henlo");
+        }
+    }
+
 }
